requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
arxiv==2.1.3
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.145
botocore==1.34.145
Brotli==1.1.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
click==8.1.7
cohere==5.6.1
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.2.0
cssselect2==0.7.0
dataclasses-json==0.6.7
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
duckduckgo_search==6.2.1
email_validator==2.2.0
emoji==2.12.1
exa-py==1.0.16
fastapi==0.111.1
fastapi-cli==0.0.4
fastavro==1.9.5
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
fireworks-ai==0.14.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
google_search_results==2.4.2
googleapis-common-protos==1.63.2
gpt-researcher==0.8.1
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
json_repair==0.25.3
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.9
langchain-anthropic==0.1.20
langchain-aws==0.1.11
langchain-chroma==0.1.2
langchain-cohere==0.1.9
langchain-community==0.2.7
langchain-core==0.2.21
langchain-experimental==0.0.62
langchain-fireworks==0.1.5
langchain-google-genai==1.0.7
langchain-google-vertexai==1.0.6
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-mistralai==0.1.10
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchain-together==0.1.4
langchain-upstage==0.1.7
langchainhub==0.1.20
langdetect==1.0.9
langgraph==0.1.9
langgraph-cli==0.1.49
langsmith==0.1.92
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.21.3
md2pdf==1.0.1
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.36.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
playwright==1.45.0
posthog==3.5.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydyf==0.11.0
pyee==11.1.0
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
pypdf==4.3.0
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyreqwest_impersonate==0.5.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tabulate==0.9.0
tavily-python==0.3.5
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.14.10
unstructured-client==0.24.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
weasyprint==62.3
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
zopfli==0.2.3
